backend/pyproject.toml
[project]
name = "agentic-trip-planner"
version = "0.1.0"
description = "Agentic LLM trip planner backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "sqlalchemy>=2.0.29",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "structlog>=24.1.0",
    "httpx>=0.27.0",
    "google-generativeai>=0.7.2",
    "tenacity>=8.2.3",
    "redis>=5.0.4",
    "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
